fastapi-boilerplate/app/schemas/health.py
from pydantic import BaseModel, Field

class HealthStatus(BaseModel):
    """
    Pydantic model for the health check endpoint response.
    """
    status: str = Field(..., description="The operational status of the service.", example="ok")
