fastapi-boilerplate/pyproject.toml
[project]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "A production-ready FastAPI boilerplate with uv and loguru."
authors = [
    { name = "Your Name", email = "you@example.com" },
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "pydantic-settings>=2.3.4",
    "loguru>=0.7.2",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-exporter-otlp>=1.25.0",
]

[project.scripts]
start_dev = "app.main:run_dev_server"
start_prod = "app.main:run_prod_server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Include the 'app' directory as a top-level package.
packages = ["app"]
